pyproject.toml
[project]
name = "flood-agent"
version = "0.1.0"
requires-python = ">=3.11"
description = "Multi-agent flood detection system using Google ADK and MCP"

dependencies = [
    # Google ADK for multi-agent orchestration
    "google-adk>=0.2.0",
    # MCP dependencies
    "fastmcp>=0.2.0",        # MCP server framework
    "mcp>=1.0.0",            # MCP client for ADK integration
    # Web server dependencies
    "fastapi>=0.104.0",      # FastAPI web framework
    "uvicorn[standard]>=0.24.0",  # ASGI server
    "websockets>=12.0",      # WebSocket support
    "python-multipart>=0.0.6",  # Form/file upload support
    "pydantic-settings>=2.0.0",  # Settings management
    # Core utilities
    "requests>=2.31.0",      # HTTP client
    "python-dotenv>=1.0.0",  # Environment variables
    "gradio-client>=0.10.0", # Prithvi HuggingFace API
    "pillow>=10.0.0",        # Image processing
    "numpy>=1.24.0",         # Array operations
    "rasterio>=1.3.0",       # Geospatial raster I/O
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcp_server", "agents"]
